Pre-compiler tupling of method returns; uniform handling of free vars in lambdas
The pre-compiler now transforms each method with multiple (non-ghost) out
parameters into one that returns a single tuple, when targeting a language
(currently Java or JavaScript) that has neither output parameters (as C# has)
nor support for returning multiple values (as Go has).

This work exposed a bug (#531) in the handling of free occurrences of out
parameters in constructs that are translated as lambdas in the target language.
Namely, such out parameters need to be copied into temporary variables in the
same way that free variables of source-level lambdas are copied.  (Previously,
this was done with let expressions but not other constructs like set
comprehensions.)  Also, Java requires more care than other languages in dealing
with free vars in lambdas, as *no* mutable variable may occur free in a lambda
in Java.  (More precisely, any free variable in a lambda or inner class must be
"effectively final," which is to say it could have been declared as final even
if it wasn't.)

The simplest solution to this was to take the way free variables are handled
for source-level lambdas (see the LambdaExpr case of TrExpr), generalize it a
bit, and use it for anything translated into a lambda: copy the free variables
into temporaries (using a beta redex), then, before translating the body,
perform substitution on the source term, replacing each free variable with its
corresponding temporary.  This makes it unnecessary to remember any state
associated with generating a lambda in the target language; in particular, the
inLetExprBody parameter isn't needed anymore.

FIXME: Currently, Dafny assumes that all tuple types that the program will need
are created by the parser, since each tuple type is created in a half-finished
state (no destructors, for instance) and needs the resolver to run to finish
it.  For the tupling pass, this means that the program needs to mention each
tuple type that will be needed in the final program (see SillyMethod() in
Arrays.dfy and Calls.dfy).  This is of course a temporary workaround; somehow
we need to be able to create tuple types in post-resolver code.